requirements.txt
seaborn
pandas
numpy
matplotlib
streamlit
langchain-groq
langchain_community
PyPDF
sentence-transformers
python-dotenv
langchain-unstructured
langchain-ollama
import-ipynb
langchain-chroma
chromadb
pymysql
sqlalchemy
google-genai
pdf2image